requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-filter==23.3
django-environ==0.11.2
django-rest-swagger==2.2.0
drf-spectacular==0.26.5
psycopg2-binary==2.9.6
